feat: create separated component for filter menu
